Rename method_body_type_checker file
